fix starting state of sum.c for symbol resolution

+++ linker/symbols/sum.c
@@ -1,3 +1,3 @@
-double globalX = 10;
+int globalX = 10;
 
 int sum(int a, int b) { return a + b + globalX; }
